% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{infer}
\alias{infer}
\title{infer}
\usage{
infer(model, intervene = NULL, observe = NULL)
}
\arguments{
\item{model}{OpenMX RAM model or lavaan model}

\item{intervene}{named vector with variables on which is intervened and the
respective intervention levels}

\item{observe}{named vector with variables for which a specific value is observed
(conditioning)}
}
\value{
list with expected means and covariances
}
\description{
Computes the interventional and / or conditional distribution of a SEM based
on an intervention and / or an observation of variable values.
}
\details{
For a thorough introduction to interventional and conditional distributions in
SEM, please see Gische et al. (2021).


References:

Gische, C., West, S. G., & Voelkle, M. C. (2021). Forecasting causal effects
of interventions versus predicting future outcomes. Structural Equation Modeling:
A Multidisciplinary Journal, 28(3), 475-492.
}
\examples{
library(mxsem)

model <- '
  # latent variable definitions
     ind60 =~ x1 + x2 + x3
     dem60 =~ y1 + a*y2 + b*y3 + c*y4
     dem65 =~ y5 + a*y6 + b*y7 + c*y8

  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60

  # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
'

fit <- mxsem(model = model,
             data  = OpenMx::Bollen) |>
  mxTryHard()

# intervention on dem60
infer(model = fit,
      intervene = c("dem60" = 2))
# condition on dem60
infer(model = fit,
      observe = c("dem60" = 2))
# condition and intervene
infer(model = fit,
      intervene = c("ind60" = -1),
      observe = c("dem60" = 2))
}
